requirements.txt
ConfigParser
Pillow
numpy
selenium
